pc64: enable iahci support

+++ sys/src/9/pc/sdiahci.c
@@ -2133,7 +2133,7 @@ static SDev*
 iapnp(void)
 {
 	int i, n, nunit, type;
-	ulong io;
+	uintptr io;
 	Ctlr *c;
 	Drive *d;
 	Pcidev *p;
